Reorganising the source files into the appropriate packages (challenge Id)
