+++ src/main/scala/befaster/solutions/SUM/Sum.scala
@@ -1,4 +1,4 @@
-package befaster.solutions
+package befaster.solutions.SUM
 
 import befaster.runner.SolutionNotImplementedException
