+++ src/test/scala/befaster/solutions/SUM/SumTest.scala
@@ -1,4 +1,4 @@
-package befaster.solutions
+package befaster.solutions.SUM
 
 import org.scalatest.{FlatSpec, Matchers}
